feat: move source code and start adding logger
