+++ tdm-be/.eslintrc
@@ -1,9 +1,12 @@
 {
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "tsconfig.json"
-  },
-  "extends": [
-    "@draconides/eslint-config-ts"
-  ]
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+        "project": "tsconfig.json"
+    },
+    "env": {
+        "node": true
+    },
+    "extends": [
+        "@draconides/eslint-config-ts"
+    ]
 }
