+++ tdm-be/package.json
@@ -18,27 +18,28 @@
   "author": "",
   "license": "CeCILL",
   "dependencies": {
-    "axios": "^1.6.2",
+    "axios": "^1.6.7",
     "cors": "^2.8.5",
+    "express": "^4.18.2",
     "express-basic-auth": "^1.2.1",
     "multer": "^1.4.5-lts.1",
     "node-cron": "^3.0.3",
-    "nodemailer": "^6.9.7",
+    "nodemailer": "^6.9.9",
     "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.0"
+    "swagger-ui-express": "^5.0.0",
+    "winston": "^3.11.0"
   },
   "devDependencies": {
     "@draconides/eslint-config-ts": "^1.2.0",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/multer": "^1.4.11",
-    "@types/node": "^18.19.14",
+    "@types/node": "^18.19.15",
     "@types/node-cron": "^3.0.11",
     "@types/nodemailer": "^6.4.14",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui-express": "^4.1.6",
     "eslint": "^8.56.0",
-    "express": "^4.18.2",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
   }
