+++ tdm-be/src/logger.ts
@@ -1,0 +1,47 @@
+import { createLogger, format, transports } from 'winston';
+import type { Logger } from 'winston';
+
+/**
+ * Create a winston logger with an associated label
+ * @param name Label use with the logger
+ */
+const create = (name: string): Logger => {
+    return createLogger({
+        format: format.combine(
+            format.label({ label: name }),
+            format.timestamp(),
+            format.printf((info) => {
+                return `${info.timestamp} [${info.label}] ${info.level}: ${info.message}`;
+            }),
+        ),
+        transports: [
+            new transports.Console(),
+            new transports.File({
+                filename: name === 'default' ? 'logs/combined.log' : `logs/${name}-combined.log`,
+            }),
+            new transports.File({
+                level: 'debug',
+                filename: name === 'default' ? 'logs/debug.log' : `logs/${name}-debug.log`,
+            }),
+        ],
+    });
+};
+
+/**
+ * Default logger, the logger can be used for general propos
+ */
+export const defaultLogger: Logger = create('default');
+/**
+ * Logger use by the express http server
+ */
+export const httpLogger: Logger = create('http');
+/**
+ * Logger use to log mail process
+ */
+export const mailLogger: Logger = create('mail');
+/**
+ * Logger use to log cron process
+ */
+export const cronLogger: Logger = create('cron');
+
+export default defaultLogger;
